feat(delete-location): validate units occupied on delete location

+++ src/main/java/ru/itmo/highload/storoom/services/LocationService.java
@@ -4,13 +4,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import ru.itmo.highload.storoom.consts.UnitStatus;
 import ru.itmo.highload.storoom.exceptions.ResourceAlreadyExistsException;
 import ru.itmo.highload.storoom.exceptions.ResourceNotFoundException;
 import ru.itmo.highload.storoom.models.DTOs;
 import ru.itmo.highload.storoom.models.LocationEntity;
 import ru.itmo.highload.storoom.models.OwnerEntity;
+import ru.itmo.highload.storoom.models.UnitEntity;
 import ru.itmo.highload.storoom.repositories.LocationRepo;
 import ru.itmo.highload.storoom.repositories.OwnerRepo;
+import ru.itmo.highload.storoom.repositories.UnitRepo;
 import ru.itmo.highload.storoom.utils.Mapper;
 
 import java.util.List;
@@ -22,6 +25,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class LocationService {
     private final LocationRepo locationRepo;
+    private final UnitRepo unitRepo;
     private final OwnerRepo ownerRepo;
 
     public LocationEntity getEntityById(UUID id) {
@@ -55,6 +59,10 @@ public class LocationService {
 
     public void delete(UUID id) {
         LocationEntity locationEntity = getEntityById(id);
+        List<UnitEntity> units = unitRepo.getAllByLocation(locationEntity);
+        if (units.stream().anyMatch(u -> u.getStatus().equals(UnitStatus.occupied))) {
+            throw new IllegalArgumentException("Close all units orders before deleting the location");
+        }
         locationRepo.delete(locationEntity);
     }
